backend/SyntheticDataset/video-gen/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
websockets>=12.0
torch>=2.4.0
torchvision>=0.19.0
diffusers>=0.32.0
transformers>=4.45.0
accelerate>=0.34.0
sentencepiece>=0.2.0
Pillow>=10.4.0
numpy>=1.26.0
imageio>=2.35.0
imageio-ffmpeg>=0.5.1
